Phase 6 : gérer les playlists.

+++ src/Repository/PlaylistRepository.php
@@ -44,7 +44,9 @@ class PlaylistRepository extends ServiceEntityRepository
 
     public function remove(Playlist $entity, bool $flush = false): void
     {
-        $this->getEntityManager()->remove($entity);
+        if($entity->getFormations()->isEmpty()){
+                $this->getEntityManager()->remove($entity);
+            }
 
         if ($flush) {
             $this->getEntityManager()->flush();
